Added RTree Dependency
Update Luau to 0.662
Added CVar Category "Database"
Improved StringUtil::StringIsNumeric
Added ClientDB Functions (TryGet, Clone, GetStringTable, HasString)
Fixed an issue in ClientDB where MaxID would desync
Added Typedef for ClientDB StringRef
Updated Model Collision Indices to 32 bit
Added New GameDefines (ObjectGuid and Data Related structures)
Added new Network Opcodes (Server_ItemCreate, Server_ContainerCreate, Client_ContainerSwapSlots, Server_ContainerAddToSlot, Server_ContainerRemoveFromSlot, Server_ContainerSwapSlots)
Added new Cheat Commands (SetItemTemplate, SetItemStatTemplate, SetItemArmorTemplate, SetItemWeaponTemplate, SetItemShieldTemplate, AddItem, RemoveItem)
Fixed an issue where Mouse Position would continue to be updated while cursor was virtualized
Improved Text Size Calculation in Font
Reduced brightness of the Default White Texture
Improved Cursor Information Availability in InputManager

+++ Source/FileFormat/FileFormat/Novus/Model/ComplexModel.cpp
@@ -328,7 +328,7 @@ namespace Model
 
                 if (numCollisionIndices > 0)
                 {
-                    output.write(reinterpret_cast<const char*>(&collisionIndices[0]), numCollisionIndices * sizeof(u16));
+                    output.write(reinterpret_cast<const char*>(&collisionIndices[0]), numCollisionIndices * sizeof(u32));
                 }
             }
